+++ pymycobot/common.py
@@ -289,6 +289,10 @@ class DataProcessor(object):
 
         # process valid data
         res = []
+        if genre in [ProtocolCode.GET_SERVO_VOLTAGES, ProtocolCode.GET_SERVO_STATUS, ProtocolCode.GET_SERVO_TEMPS]:
+            for i in valid_data:
+                res.append(i)
+            return res
         if data_len in [6, 8, 12, 14, 24, 60]:
             for header_i in range(0, len(valid_data), 2):
                 one = valid_data[header_i : header_i + 2]