+++ pymycobot/mybuddy.py
@@ -106,9 +106,13 @@ class MyBuddy(MyBuddyCommandGenerator):
                             datas = b'\xfe'
                             pre = k  
             except:
+                self.log.debug("_read: {}".format(datas))
+                
                 datas = None
                 break
         else:
+            self.log.debug("_read: {}".format(datas))
+
             datas = None
         self.log.debug("_read: {}".format(datas))
         return datas
@@ -155,7 +159,9 @@ class MyBuddy(MyBuddyCommandGenerator):
                 ProtocolCode.GET_MOVEMENT_TYPE,
                 ProtocolCode.GET_REFERENCE_FRAME,
                 ProtocolCode.GET_JOINT_MIN_ANGLE,
-                ProtocolCode.GET_JOINT_MAX_ANGLE
+                ProtocolCode.GET_JOINT_MAX_ANGLE,
+                ProtocolCode.GET_FRESH_MODE,
+                # ProtocolCode.GET_SERVO_CURRENTS
             ]:
                 return self._process_single(res)
             elif genre in [ProtocolCode.GET_ANGLES]: